added gameroom component, game-room.css and API_BASE_URL_SOCKET to config.js
