+++ src/App.js
@@ -40,9 +40,8 @@ export class App extends React.Component {
   render() {
     return (
       <div className="app">
-          <HeaderBar />
-          <Route exact path="/" component={LandingPage} />
-
+        <HeaderBar />
+        <Route exact path="/" component={LandingPage} />
         <Route exact path="/dashboard" component={Dashboard} />
         <Route exact path="/register" component={RegisterPage} />
       </div>
