+++ src/components/game-room.css
@@ -1,0 +1,5 @@
+.challenger-typing {
+  height: 100px;
+  width: 200px;
+  border: 1px solid black;
+}
