+++ src/config.js
@@ -1,2 +1,3 @@
 export const API_BASE_URL =
     process.env.REACT_APP_API_BASE_URL || 'http://localhost:8080/api';
+export const API_BASE_URL_SOCKET = process.env.REACT_APP_API_BASE_URL || 'http://localhost:8080';
